+++ code/main.cpp
@@ -1,7 +1,7 @@
 #include <QApplication>
 #include <iostream>
 
-#include "mainwindow.h"
+#include "./window/mainwindow.h"
 
 int main(int argc, char *argv[]) {
     QApplication a(argc, argv);
